feat: 버그 수정 및 리펙토링

+++ DungeonFarming/DungeonFarming/DataBaseServices/GameDb/ItemBundle.cs
@@ -2,7 +2,7 @@
 {
     public class ItemBundle
     {
-        public Int64 itemCode { get; set; }
+        public Int32 itemCode { get; set; }
         public Int64 itemCount { get; set; }
     }
 }
